[common] Move common/nodeutil -> util/nodeutil
- common package no longer exists
